+++ changelog.md
@@ -3,6 +3,7 @@ For the latest version of this document, please see [https://github.com/bos/aeso
 ### 1.5.2.0
 
 * Add `Ord Value` instance
+* Export `rejectUnknownFields` from `Data.Aeson`
 
 ### 1.5.1.0